requirements.txt
python-telegram-bot[ext]==21.5
aiosqlite==0.20.0
aiohttp==3.9.5
